feat(odyssey-react): add withStyles to Tab, refactor implementaion to align with other components
